Fix #189 Separate Elements from Pages * Separate elements from pages. Created new files to create the elements: lib/creature.tsx lib/waves.tsx lib/vocabulary-render.tsx lib/debug.tsx * Pared down -page files to just render the elements
