+++ lib/pages/mandala-page.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import { Page } from "../page";
+import { Mandala } from "./mandala";
+
+export const MandalaPage: React.FC<{}> = () => (
+  <Page title="Mandala!">
+    <Mandala />
+  </Page>
+);
