+++ lib/pages/mandala/mandala-design.avsc.json
