+++ lib/pages/mandala/mandala-design.v1.avsc.json
